+++ R/RcppExports.R
@@ -1,12 +1,8 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-testing_resize_array <- function(size, values) {
-    invisible(.Call(`_scPipe_testing_resize_array`, size, values))
-}
-
-check_barcode_reads <- function(fastq, barcodeseqs, barcode_start, barcode_length, lines_to_search) {
-    .Call(`_scPipe_check_barcode_reads`, fastq, barcodeseqs, barcode_start, barcode_length, lines_to_search)
+check_barcode_reads <- function(fastq, barcodeseqs, barcode_start, barcode_length, lines_to_search, threshold) {
+    .Call(`_scPipe_check_barcode_reads`, fastq, barcodeseqs, barcode_start, barcode_length, lines_to_search, threshold)
 }
 
 rcpp_sc_trim_barcode_paired <- function(outfq, r1, r2, bs1, bl1, bs2, bl2, us, ul, rmlow, rmN, minq, numbq, write_gz) {
