+++ R/sc_atac_trim_barcode.R
@@ -1,25 +1,25 @@
 #' sc_atac_trim_barcode()
 #'
-#' @return 
+#' @return
 #'
 #' @examples
 #' \dontrun{
-#' 
-#' 
+#'
+#'
 #' }
 #'
 #' @export
 #'
 sc_atac_trim_barcode = function(
-  r1, 
-  r2, 
-  bc_file, 
-  output_folder = "", 
-  bc_start=-1, 
-  bc_length=-1, 
-  umi_start=0, 
-  umi_length=0, 
-  umi_in = "both", 
+  r1,
+  r2,
+  bc_file,
+  output_folder = "",
+  bc_start=-1,
+  bc_length=-1,
+  umi_start=0,
+  umi_length=0,
+  umi_in = "both",
   rmN = FALSE,
   rmlow = FALSE,
   min_qual = 20,
@@ -28,7 +28,7 @@ sc_atac_trim_barcode = function(
   id1_len = -1,
   id2_st = -1,
   id2_len = -10) {
-  
+
   if(output_folder == ''){
     output_folder = file.path(getwd(), "scPipe-atac-output")
   }
@@ -36,30 +36,30 @@ sc_atac_trim_barcode = function(
     dir.create(output_folder,recursive=TRUE)
     cat("Output Directory Does Not Exist. Created Directory: ", output_folder, "\n")
   }
-  
-  
+
+
   log_and_stats_folder = paste0(output_folder, "/log_and_stats/")
   dir.create(log_and_stats_folder, showWarnings = F)
   log_file = paste0(log_and_stats_folder, "log_file.txt")
   stats_file = paste0(log_and_stats_folder, "stats_file_trimbarcode.txt")
   if(!file.exists(log_file)) file.create(log_file)
   file.create(stats_file)
-  
+
   cat(
     paste0(
       "trimbarcode starts at ",
       as.character(Sys.time()),
       "\n"
-    ), 
+    ),
     file = log_file, append = T)
-  
+
   if (substr(r1, nchar(r1) - 2, nchar(r1)) == ".gz") {
     write_gz = TRUE
   }
   else {
     write_gz = FALSE
   }
-  
+
   if (!is.null(bc_file)) {
     if (!file.exists(r1)) {stop("read1 fastq file does not exist.")}
     i=1;
@@ -68,7 +68,7 @@ sc_atac_trim_barcode = function(
       bc_file[i] = path.expand(bc)
       i = i+1;
     }
-    
+
     if(umi_start != 0){
       if(umi_in %in% c("both", "R1", "R2")){
         cat("UMI Present in: ", umi_in, "\n")
@@ -76,10 +76,10 @@ sc_atac_trim_barcode = function(
         stop("Invalid value of umi_in. Possible values are both, R1 and R2")
       }
     }
-    
+
     # expand tilde to home path for downstream gzopen() call
     r1 = path.expand(r1)
-    
+
     if(!is.null(r2)){
       if (!file.exists(r2)) {stop("read2 file does not exist.")}
       r2 = path.expand(r2)
@@ -89,11 +89,11 @@ sc_atac_trim_barcode = function(
     cat("Saving the output at location: ")
     cat(output_folder)
     cat("\n")
-    
+
     if(file_ext(bc_file) != 'csv'){
       out_vec = rcpp_sc_atac_trim_barcode_paired(
-        output_folder, 
-        r1, 
+        output_folder,
+        r1,
         bc_file,
         r2,write_gz,
         rmN,
@@ -104,15 +104,15 @@ sc_atac_trim_barcode = function(
         id1_len,
         id2_st,
         id2_len,
-        umi_start, 
+        umi_start,
         umi_length)
-      
-      cat("Total Reads: ", out_vec[1], 
-          "\nTotal N's removed: ", out_vec[2], 
-          "\nremoved_low_qual: ", out_vec[3], 
+
+      cat("Total Reads: ", out_vec[1],
+          "\nTotal N's removed: ", out_vec[2],
+          "\nremoved_low_qual: ", out_vec[3],
           "\n",
           file = stats_file, append = T)
-      
+
     }
     else {
       cat("Using barcode CSV file, since barcode FastQ file is not passed \n")
@@ -120,22 +120,21 @@ sc_atac_trim_barcode = function(
         stop("Please pass bc_start and bc_length values")
       }
 
-      # this only works for .fastq files, not .fastq.gz Should be fixed for .gz files.
-      # this should be where we do the subset check for correct bc_start positions.
-      if (!check_barcode_start_position(r1, bc_file, bc_start, bc_length, 10000)) {
+      # Check if given barcode start position is valid
+      if (!check_barcode_start_position(r1, bc_file, bc_start, bc_length, 10000, .8)) {
           stop("Please change bc_start and try again")
       }
 
       out_vec = rcpp_sc_atac_trim_barcode(
-        output_folder, 
-        r1, 
-        r2, 
-        bc_file, 
-        bc_start, 
-        bc_length, 
-        umi_start, 
-        umi_length, 
-        umi_in, 
+        output_folder,
+        r1,
+        r2,
+        bc_file,
+        bc_start,
+        bc_length,
+        umi_start,
+        umi_length,
+        umi_in,
         write_gz,
         rmN,
         rmlow,
@@ -145,29 +144,29 @@ sc_atac_trim_barcode = function(
         id1_len,
         id2_st,
         id2_len)
-      
-      cat("Total Reads: ", out_vec[1], 
-          "\nTotal N's removed: ", out_vec[2], 
-          "\nremoved_low_qual: ", out_vec[3], 
-          "\nExact match Reads: ", out_vec[4], 
-          "\nApprox Match Reads: ", out_vec[5], 
+
+      cat("Total Reads: ", out_vec[1],
+          "\nTotal N's removed: ", out_vec[2],
+          "\nremoved_low_qual: ", out_vec[3],
+          "\nExact match Reads: ", out_vec[4],
+          "\nApprox Match Reads: ", out_vec[5],
           "\n",
           file = stats_file, append = T)
-      
+
     }
   }else{
     stop("Barcode file is mandatory")
   }
-  
-  
+
+
   cat(
     paste0(
       "trimbarcode finishes at ",
       as.character(Sys.time()),
       "\n\n"
-    ), 
+    ),
     file = log_file, append = T)
-  
+
   return(out_vec)
 }
