+++ scPipe-atac-output/log_and_stats/log_file.txt
@@ -1,0 +1,1 @@
+trimbarcode starts at 2021-01-26 12:15:53
